JCL-337: Use verify endpoint in tests (#527)
Call verify to check the validity of the issued grant, and the invalidity of the revoked grant.

The mocked VC service used for local integration test is added a `/verify` endoint. It is stateful: an access grant is considered valid until a revocation request has been made, and it is then considered invalid.