Remove inner verification status class
Instead of having a mutable verification status class (which was intended to be private, but JSONB complained about that), the public AccessCredentialVerification class has now public setters and a default constructor to be usable directly by the JSON binding service. The setters will make the attributes immutable, so that they can only be set once.