Update the UMA module to use UmaMetadata from api
It is not possible to remove the Metadata class internal to the uma module entirely from the API, because that would be a breaking change for UmaClient::metadata (a TODO has been added to this effect). Other than that, this enforces that the UMA module codes to the UmaMetadata interface rather than the Metadata class.