JCL-403: SolidClient builds ProblemDetails
`SolidClient` now parses HTTP error responses if they are compliant with
RFC9457 in order to throw a structured exception. In case when no JSON
parser is available in the classpath, the default behavior is to build
a ProblemDetails object that only relies on the data available in the
HTTP response status code.