JCL-402: Data classes for HTTP Problem Details support (#1160)
This creates the `ProblemDetails` class, which represents the content of an RFC9457-compliant HTTP error response. `ProblemDetails` encapsulates the SPI lookup required to discover a JSON service used to deserialize RFC9457 error responses. If no JSON parser is available, `ProblemDetails` falls back to providing a minimal object with all its fields nulled except for the type set to a default value and the error code set to the value associated with the response. Unlike what was suggested in the initial design, no default is applied to the title based on the HTTP status phrase, because not all the HTTP client libraries we support expose sur HTTP status phrase.

The tests for `ProblemDetails` parsing are in the solid-client module, because it will need the additional json service anyways in the test dependencies for testing the rest of this feature, and putting these tests in the api module creates a circular dependency.

The `Response` interface from the `api` module now also exposes a new `BodyHandler` that will throw on HTTP error. This body handler is composable with another body handler that handles success cases, and projects the HTTP response to the desirable object. Both the exception mapper and the success criterion are configurable. This new body handler throws a `ClientHttpException`, which is a new Exception type parent of `SolidClientException`. The HTTP-specific members of `SolidClientException` have been lifted up to `ClientHttpException`.