JCL-402: Low-level client throws enriched exceptions (#1157)
`SolidClient` now uses the `throwOnError` body mapper to handle error responses, and provides a custom exception mapper to throw the appropriate specialized exception on HTTP error response. Overall, the behavior of the library doesn't change, the only addition is that the thrown Exception object now has a `ProblemDetails` instance attached to it.