JCL-402: RDF4J body handlers http error handling (#1162)
This deprecates the current methods in `RDF4JBodyHandlers`, and replaces them with similar method throwing an appropriate exception with error details on HTTP error instead of returning an empty dataset. The new method now have `RDF4J` in their name for this not to be a breaking change: `ofModel` becomes `ofRDF4JModel`, etc.